+++ src/components/NoSmokingList.js
@@ -1,24 +1,26 @@
 import React from 'react';
-import '../../styles/SmokingArea.css';
+import '../styles/SmokingArea.css';
 
-function SmokingList({smokingAreas, onListClick}) {
+function SmokingList({nosmokingAreas, onListClick}) {
     return(
       <div className="smoking-areas">
-          {smokingAreas.length > 0 ? (
-              smokingAreas.map((area, index) => (
+          {nosmokingAreas.length > 0 ? (
+              nosmokingAreas.map((area, index) => (
                   <div 
                     key={index}
                     className="facility-card"
                     onClick={() => onListClick(index)}
                     style={{ cursor: "pointer" }}
                   >
-                    <p>📍 {area.address} {area.detail}</p>
-                    <p>🚩 {area.open}</p>
+                    <p>📍 {area.address}</p>
+                    <p>범위: {area.scope}</p>
+                    <p>🧾 위반 과태료: {Number(area.fine).toLocaleString()} 원</p>
+                    <p>지정 근거명: {area.statute}</p>
                   </div>
                 ))
               ) : (
                 <div className="facility-info">
-                  <p>No smoking areas found.</p>
+                  <p>No No-smoking areas found.</p>
                 </div>
             )}
       </div>
