+++ src/pages/ParkList.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
-import DropDown from "../components/park/DropDown";
-import ParkCard from "../components/park/ParkCard";
-import KakaoMap from "../components/park/KakaoMap";
+import DropDown from "../components/DropDown";
+import ParkCard from "../components/ParkCard";
+import KakaoMap from "../components/KakaoMap";
 import "../styles/ParkList.css";
-import api from "../api/api"; // <-- axios 인스턴스 import
+import api from "../api"; // <-- axios 인스턴스 import
 
 const ParkList = () => {
   // ----- 주 상태들 -----
