+++ src/pages/Signup.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import api from "../api/api";
+import api from "../api";
 import "../styles/Signup.css";
 
 function Signup() {
